Refactor: Factor out `Resolution`
